requirements.txt
alembic==1.4.2
certifi==2019.11.28
cftime==1.1.1.2
chardet==3.0.4
click==7.1.1
decorator==4.4.2
et-xmlfile==1.0.1
Flask==1.1.1
Flask-GraphQL==2.0.1
flask-marshmallow==0.11.0
Flask-Migrate==2.5.3
flask-oidc==1.4.0
Flask-SQLAlchemy==2.4.1
future==0.18.2
geocoder==1.38.1
geographiclib==1.50
geopy==1.21.0
googlemaps==4.2.0
graphql-core==2.3.2
graphql-server-core==1.2.0
httplib2==0.17.0
idna==2.9
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.1
joblib==0.16.0
Mako==1.1.2
MarkupSafe==1.1.1
marshmallow==3.5.1
marshmallow-sqlalchemy==0.22.3
netCDF4==1.5.3
nltk==3.5
numpy==1.18.2
oauth2client==4.1.3
openpyxl==3.0.3
pandas==1.0.3
pbkdf2==1.3
pgeocode==0.2.1
Pillow==7.0.0
promise==2.3
psycopg2-binary==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
python-docx==0.8.10
python-editor==1.0.4
pytz==2019.3
pyjwt==1.7.1
ratelim==0.1.6
regex==2020.6.8
requests==2.23.0
require==0.1.1
requires.io==0.2.6
rsa==4.0
Rx==1.6.1
six==1.14.0
SQLAlchemy==1.3.15
tqdm==4.47.0
urllib3==1.25.8
waitress==1.4.3
Werkzeug==0.16.1
XlsxWriter==1.2.9
xmljson==0.2.0
